maqueta front terminada, hecha componente y full responsive
