+++ src/components/MangasTitle/mangasTitle.css
@@ -1,0 +1,41 @@
+.mangas .title{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 6.33vh;
+}
+
+.mangas .title>h2{
+    font-weight: 700;
+    font-size: 40px;
+    line-height: 95.19%;
+    color: #FFFFFF;
+}
+
+.mangas .title .search-bar{
+    display: flex;
+    align-items: center;
+    padding: 26px;
+    gap: 2.32vw;
+    width: 91.395vw;
+    height: 8.154vh;
+    background: #FFFFFF;
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    border-radius: 80px;
+}
+
+.mangas .title .search-bar img{
+    height: 37px;
+    width: 37px;
+}
+
+.mangas .title .search-bar input{
+    border: none;
+    outline: none;
+    font-weight: 400;
+    font-size: 24px;
+    line-height: 95.19%;
+    color: #9795A4;
+    overflow: hidden;
+    width: 100%;
+}
